removing unneccesary rounding

+++ core/src/com/faust/doomlike/utils/SectorWrapper.java
@@ -20,7 +20,7 @@ public class SectorWrapper {
     }
 
     private final SectorData sectorData;
-    private final Map<Integer,Float> surfaceYforXMap =new HashMap<>();
+    private final Map<Float,Float> surfaceYforXMap =new HashMap<>();
     private SurfaceShownEnum surfaceToShow = SurfaceShownEnum.NONE;
     private float depth = 0;
 
@@ -28,7 +28,7 @@ public class SectorWrapper {
         this.sectorData = sectorData;
     }
 
-    public Map<Integer,Float>  getSurfaceYforXMap() {
+    public Map<Float,Float>  getSurfaceYforXMap() {
         return surfaceYforXMap;
     }
